Support for embedding Pig scripts in Java test files
Uses a neat trick presented by Adrian Walker in this blog post: http://www.adrianwalker.org/2011/12/java-multiline-string.html

With this the pig scripts can be embedded directly in the Java files as javadoc comments on class fields.  Tests are much easier to work with this way.

I removed all pig test files for the util tests to demonstrate.

Also added an ant target "eclipse" which generates some eclipse files.  It figures out what the paths should be based on the environment.  This fixes an issue where paths would be incorrect when working in another repo.