Fix to correctly load schema for input paths
The map from input path to schema was not set up correctly in the record reader, so the schema wasn't be loaded as a result and instead was null.  This corrects the problem and also adds verification that the schema can be determined.

This also fixes an issue where only schemas for the most recent day of input would be loaded.  This could cause a problems for joins as data for the most recent day may not be available across all inputs.  As a result these missing inputs did not have schemas loaded.