Clean up test classpaths
Some JARs, like fastutil, are packaged in the datafu JAR under a different namespace.  For this reason they are not included in the pom.  To properly test datafu these JARs should not be in the namespace used for testing.  This wasn't the case.  So we couldn't be sure that the datafu JAR worked without the fastutil JAR be downloaded.  I've fixed this by having separate classpaths for building the JAR, building the tests, and running the tests.

Other changes:
* Updated NOTICE.
* Updated pom-template.
* TransposeTupleToBag doc fix.