Package small subset of google's guava in DataFu
I wanted to remove Guava as a dependency in our pom.  This makes DataFu easier to use.  We are not using very many significant parts of Guava.  So, I tried autojarring it, which packages Guava in DataFu and includes only what we need.  This ended up pulling in a lot of Guava, as we were using ImmutableList in several places.  I replaced these calls to Arrays, which ships with Java.  Now there is just a handful of classes we pull in from Guava.  The file size is not significantly changed as a result.